Chapter13/requirements.txt
mysqlclient
click
blinker
redis
Flask==0.12.4
Flask-Login>0.3.0
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.1.1
Flask-WTF==0.14.2
flask-restful
flask-jwt-extended
flask-debugtoolbar
flask-admin
Flask-Caching
Flask-Babel==0.11.1
Flask-Mail
Faker==0.8.13
Flask-Assets
cssmin
jsmin
flask-bcrypt
flask-openid
flask-dance
celery
flask-celery-helper
selenium
python-dotenv

